+++ src/main/java/ru/otus/library/mappers/AuthorMapper.java
@@ -9,7 +9,7 @@ import java.sql.SQLException;
 public class AuthorMapper implements RowMapper<Author> {
     @Override
     public Author mapRow(ResultSet resultSet, int i) throws SQLException {
-        int id = resultSet.getInt("id");
+        Long id = resultSet.getLong("id");
         String surname = resultSet.getString("surname");
         String name = resultSet.getString("name");
         return new Author(id, surname, name);
