+++ src/main/java/ru/otus/library/mappers/GenreMapper.java
@@ -9,7 +9,7 @@ import java.sql.SQLException;
 public class GenreMapper implements RowMapper<Genre> {
     @Override
     public Genre mapRow(ResultSet resultSet, int i) throws SQLException {
-        int id = resultSet.getInt("id");
+        Long id = resultSet.getLong("id");
         String name = resultSet.getString("name");
         return new Genre(id, name);
     }
